+++ src/main/java/com/owncloud/android/ui/preview/PreviewTextFileFragment.java
@@ -274,7 +274,6 @@ public class PreviewTextFileFragment extends PreviewTextFragment {
                 containerActivity,
                 getActivity(),
                 false,
-                deviceInfo,
                 user
             );
             mf.filter(menu, true);